fix ui customer, listOrder
