+++ OoadProject.Core/ViewModels/Customers/Validations/NameValidation.cs
@@ -1,0 +1,23 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Globalization;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+using System.Windows.Controls;
+
+namespace OoadProject.Core.ViewModels.Customers
+{
+    public class NameValidation : ValidationRule
+    {
+        public override ValidationResult Validate(object value, CultureInfo cultureInfo)
+        {
+            string name = (string)value;
+            if (value == null || name.Trim().Length == 0)
+            {
+                return new ValidationResult(false, "Vui lòng nhập tên khách hàng");
+            }
+            return ValidationResult.ValidResult;
+        }
+    }
+}
