+++ OoadProject.Data/OoadProject.Data.csproj
@@ -62,9 +62,9 @@
     <Compile Include="Entity\AppUser\Role.cs" />
     <Compile Include="Entity\AppUser\User.cs" />
     <Compile Include="Entity\Others\Parameter.cs" />
-    <Compile Include="Migrations\202106020908533_Initial.cs" />
-    <Compile Include="Migrations\202106020908533_Initial.designer.cs">
-      <DependentUpon>202106020908533_Initial.cs</DependentUpon>
+    <Compile Include="Migrations\202106021716135_Initial.cs" />
+    <Compile Include="Migrations\202106021716135_Initial.designer.cs">
+      <DependentUpon>202106021716135_Initial.cs</DependentUpon>
     </Compile>
     <Compile Include="Migrations\Configuration.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
@@ -100,8 +100,8 @@
     <Content Remove="obj\**\*.cs" />
     <Compile Remove="obj\**\*.cs" />
     <EmbeddedResource Remove="obj\**\*.cs" />
-    <EmbeddedResource Include="Migrations\202106020908533_Initial.resx">
-      <DependentUpon>202106020908533_Initial.cs</DependentUpon>
+    <EmbeddedResource Include="Migrations\202106021716135_Initial.resx">
+      <DependentUpon>202106021716135_Initial.cs</DependentUpon>
     </EmbeddedResource>
     <None Remove="obj\**\*.cs" />
   </ItemGroup>
