--- OoadProject/View/ListImportedItemUserControl.xaml
@@ -1,122 +1,0 @@
-﻿<UserControl x:Class="OoadProject.View.ListImportedItemUserControl"
-             xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
-             xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-             xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006" 
-             xmlns:d="http://schemas.microsoft.com/expression/blend/2008" 
-             xmlns:local="clr-namespace:OoadProject.View" 
-             xmlns:materialDesign="http://materialdesigninxaml.net/winfx/xaml/themes"
-             mc:Ignorable="d" 
-             d:DesignHeight="450" d:DesignWidth="800"
-             DataContext="{StaticResource ReceiptViewModel}">
-    <Grid>
-        <Grid.ColumnDefinitions>
-            <ColumnDefinition Width="5*"></ColumnDefinition>
-            <ColumnDefinition Width="4*"></ColumnDefinition>
-        </Grid.ColumnDefinitions>
-        <Grid.RowDefinitions>
-            <RowDefinition Height="*"></RowDefinition>
-            <RowDefinition Height="9*"></RowDefinition>
-        </Grid.RowDefinitions>
-        <StackPanel Orientation="Horizontal" HorizontalAlignment="Center" VerticalAlignment="Center" >
-            <DatePicker
-                SelectedDate="{Binding DateFrom}"
-                Width="100"
-                Margin="40 0 0 0"
-                VerticalAlignment="Center"
-                materialDesign:HintAssist.Hint="Từ ngày"
-                Style="{StaticResource MaterialDesignFloatingHintDatePicker}" />
-            <materialDesign:PackIcon Kind="ChevronRight" Width="100" Height="30" VerticalAlignment="Bottom"/>
-            <DatePicker
-                SelectedDate="{Binding DateTo}"
-                Width="100"
-                materialDesign:HintAssist.Hint="Đến ngày"
-                Style="{StaticResource MaterialDesignFloatingHintDatePicker}" />
-            <Button
-                Command="{Binding SearchWithFilter}"
-                Style="{StaticResource MaterialDesignIconForegroundButton}"
-                ToolTip="Tra cứu"
-                Background="{DynamicResource MaterialDesignTextFieldBoxBackground}"
-                Margin="20 0 0 0"
-                VerticalAlignment="Bottom">
-                <materialDesign:PackIcon Kind="Magnify" Width="16" Height="16"/>
-            </Button>
-
-        </StackPanel>
-        <Grid Grid.Row="1" Grid.Column="0">
-            <Grid.RowDefinitions>
-                <RowDefinition Height="10*"></RowDefinition>
-                <RowDefinition Height="*"></RowDefinition>
-            </Grid.RowDefinitions>
-            <materialDesign:Card Grid.Row="0" Background="#FFE3F0FF"  Margin="0,5,2,0" BorderThickness="0" VerticalAlignment="Stretch" Grid.ColumnSpan="2" Grid.RowSpan="2">
-                <ScrollViewer PreviewMouseWheel="ScrollViewer_PreviewMouseWheel" VerticalScrollBarVisibility="Auto">
-                    <ListView x:Name="ListOrder" ItemsSource="{Binding Receipts}" SelectedItem="{Binding SelectedReceipt}">
-                        <ListView.ItemContainerStyle>
-                            <Style TargetType="ListViewItem">
-                                <Setter Property="HorizontalContentAlignment" Value="Center"/>
-                                <Setter Property="BorderThickness" Value="0"/>
-                                <Setter Property="Margin" Value="0 0 0 0"/>
-                                <Setter Property="Padding" Value="0"/>
-                            </Style>
-                        </ListView.ItemContainerStyle>
-                        <ListView.View >
-                            <GridView>
-                                <GridView.ColumnHeaderContainerStyle>
-                                    <Style TargetType="GridViewColumnHeader">
-                                        <Setter Property="Margin" Value="0"/>
-                                        <Setter Property="Padding" Value="0"/>
-                                        <Setter Property="Background" Value="#FFE3F0FF"/>
-                                        <Setter Property="Height" Value="40"/>
-                                    </Style>
-                                </GridView.ColumnHeaderContainerStyle>
-                                <GridViewColumn Header="Mã ĐH" DisplayMemberBinding="{Binding Id}" ></GridViewColumn>
-                                <GridViewColumn Header="Tạo lúc" DisplayMemberBinding="{Binding CreationTime}"></GridViewColumn>
-                                <GridViewColumn Header="Người tạo" DisplayMemberBinding="{Binding CreationUser}"></GridViewColumn>
-                                <GridViewColumn Header="Nhà cung cấp" DisplayMemberBinding="{Binding ProviderName}"></GridViewColumn>
-                                <GridViewColumn Header="Tổng giá trị" DisplayMemberBinding="{Binding Total}"></GridViewColumn>
-                            </GridView>
-                        </ListView.View>
-                    </ListView>
-                </ScrollViewer>
-            </materialDesign:Card>
-        </Grid>
-
-        <TextBlock Grid.Row="0" Grid.Column="1" Margin="10 0 0 0" Style="{StaticResource MaterialDesignBody2TextBlock}" VerticalAlignment="Center">
-            Chi tiết nhập hàng
-        </TextBlock>
-
-
-            <materialDesign:Card Grid.Row="1" Grid.Column="1" Background="#FFE3F0FF"  Margin="10 5 2 0" BorderThickness="0" VerticalAlignment="Stretch" Grid.ColumnSpan="3">
-            <ScrollViewer PreviewMouseWheel="ScrollViewer_PreviewMouseWheel" VerticalScrollBarVisibility="Auto">
-                <ListView x:Name="ListDetailOrder" ItemsSource="{Binding ReceiptProducts}">
-                    <ListView.ItemContainerStyle>
-                        <Style TargetType="ListViewItem">
-                            <Setter Property="HorizontalContentAlignment" Value="Center"/>
-                            <Setter Property="BorderThickness" Value="0"/>
-                            <Setter Property="Margin" Value="0 0 0 0"/>
-                            <Setter Property="Padding" Value="0"/>
-                        </Style>
-                    </ListView.ItemContainerStyle>
-                    <ListView.View >
-                        <GridView>
-                            <GridView.ColumnHeaderContainerStyle>
-                                <Style TargetType="GridViewColumnHeader">
-                                    <Setter Property="Margin" Value="0"/>
-                                    <Setter Property="Padding" Value="0"/>
-                                    <Setter Property="Background" Value="#FFE3F0FF"/>
-                                </Style>
-                            </GridView.ColumnHeaderContainerStyle>
-                            <GridViewColumn Header="Mã SP" DisplayMemberBinding="{Binding Id}"></GridViewColumn>
-                            <GridViewColumn Header="Tên SP" DisplayMemberBinding="{Binding Name}"></GridViewColumn>
-                            <GridViewColumn Header="Loại SP" DisplayMemberBinding="{Binding CategoryName}" ></GridViewColumn>
-                            <GridViewColumn Header="Số lượng" DisplayMemberBinding="{Binding Number}"></GridViewColumn>
-                            <GridViewColumn Header="Giá nhập" DisplayMemberBinding="{Binding PriceIn}"></GridViewColumn>
-                        </GridView>
-                    </ListView.View>
-                </ListView>
-            </ScrollViewer>
-        </materialDesign:Card>
-
-    </Grid>
-</UserControl>
-
-
