--- OoadProject/View/ListOrderUserControl.xaml.cs
@@ -1,122 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-using System.Windows;
-using System.Windows.Controls;
-using System.Windows.Data;
-using System.Windows.Documents;
-using System.Windows.Input;
-using System.Windows.Media;
-using System.Windows.Media.Imaging;
-using System.Windows.Navigation;
-using System.Windows.Shapes;
-
-namespace OoadProject.View
-{
-    /// <summary>
-    /// Interaction logic for ListOrderUserControl.xaml
-    /// </summary>
-    public partial class ListOrderUserControl : UserControl
-    {
-        public ListOrderUserControl()
-        {
-            InitializeComponent();
-        }
-
-        private void ScrollViewer_PreviewMouseWheel(object sender, MouseWheelEventArgs e)
-        {
-            ScrollViewer scv = (ScrollViewer)sender;
-            scv.ScrollToVerticalOffset(scv.VerticalOffset - e.Delta);
-            e.Handled = true;
-        }
-
-        private void btnAddOrder_Click(object sender, RoutedEventArgs e)
-        {
-            var command = ((Button)sender).Command;
-            if (command.CanExecute(null))
-            {
-                command.Execute(true);
-                var window = new AddOrderWindow();
-                window.ShowDialog();
-            }
-            
-        }
-
-
-        private void itemWaitForSent_Click(object sender, RoutedEventArgs e)
-        {
-            var command = itemWaitForSent.Command;
-            if (command.CanExecute(null))
-                command.Execute(0);
-        }
-
-        private void itemSent_Click(object sender, RoutedEventArgs e)
-        {
-            var command = itemSent.Command;
-            if (command.CanExecute(null))
-                command.Execute(1);
-        }
-
-        private void itemDone_Click(object sender, RoutedEventArgs e)
-        {
-            var command = itemDone.Command;
-            if (command.CanExecute(null))
-                command.Execute(2);
-        }
-
-        private void menuItemChangeStatus(object sender, RoutedEventArgs e)
-        {
-            var command = ((MenuItem)sender).Command;
-            if (command.CanExecute(null))
-            {
-                command.Execute(true);
-                btnReloadReceiptUI.Command.Execute(null);
-            }
-        }
-
-        private void menuItemUpdate_Click(object sender, RoutedEventArgs e)
-        {
-            var command = ((MenuItem)sender).Command;
-            if (command.CanExecute(null))
-            {
-                int id;
-                if (Int32.TryParse(tbSelectedId.Text, out id))
-                {
-                    try
-                    {
-                        command.Execute(id);
-
-                        EditOrderWindow w = new EditOrderWindow();
-                        w.ShowDialog();
-                    }
-                    catch (Exception ex)
-                    {
-                        MessageBox.Show(ex.Message, "Không thể thực hiện thao tác này", MessageBoxButton.OK, MessageBoxImage.Error);
-                    }
-                }
-            }
-        }
-
-        private void menuItemDelete_Click(object sender, RoutedEventArgs e)
-        {
-            var result = MessageBox.Show("Xác nhận xóa đơn đặt hàng?", "Xác nhận", MessageBoxButton.OKCancel, MessageBoxImage.Question);
-            var command = ((MenuItem)sender).Command;
-
-            if (result == MessageBoxResult.OK && command.CanExecute(null))
-            {
-                try
-                {
-                    command.Execute(true);
-
-                    MessageBox.Show("Xóa thành công!");
-                }
-                catch (Exception ex)
-                {
-                    MessageBox.Show(ex.Message, "Không thể thực hiện thao tác này", MessageBoxButton.OK, MessageBoxImage.Error);
-                }
-            }
-        }
-    }
-}
