+++ OoadProject/View/ManagementOrderUserControl.xaml.cs
@@ -25,61 +25,115 @@ namespace OoadProject.View
             InitializeComponent();
         }
 
-        private void btnDatHang_GotFocus(object sender, RoutedEventArgs e)
+        private void btnAddOrder_Click(object sender, RoutedEventArgs e)
         {
-            btnDatHang.Opacity = 1;
-            btnNhapHang.Opacity = 0.8;
-            btnDSNhapHang.Opacity = 0.8;
-        }
-
-        private void btnNhapHang_GotFocus(object sender, RoutedEventArgs e)
-        {
-            btnDatHang.Opacity = 0.8;
-            btnNhapHang.Opacity = 1;
-            btnDSNhapHang.Opacity = 0.8;
-        }
-
-        private void btnDSNhapHang_GotFocus(object sender, RoutedEventArgs e)
-        {
-            btnDatHang.Opacity = 0.8;
-            btnNhapHang.Opacity = 0.8;
-            btnDSNhapHang.Opacity = 1;
-        }
-
-        private void btnDatHang_Click(object sender, RoutedEventArgs e)
-        {
-            if (viewControl.Name != "ListOrderUserControl")
+            var command = ((Button)sender).Command;
+            if (command.CanExecute(null))
             {
-                viewControl.Content = new ListOrderUserControl();
-                viewControl.Name = "ListOrderUserControl";
+                command.Execute(true);
+                var window = new AddOrderWindow();
+                window.ShowDialog();
+            }
+
+        }
+
+        private void itemWaitForSent_Click(object sender, RoutedEventArgs e)
+        {
+            var command = itemWaitForSent.Command;
+            if (command.CanExecute(null))
+                command.Execute(0);
+        }
+
+        private void itemSent_Click(object sender, RoutedEventArgs e)
+        {
+            var command = itemSent.Command;
+            if (command.CanExecute(null))
+                command.Execute(1);
+        }
+
+        private void itemDone_Click(object sender, RoutedEventArgs e)
+        {
+            var command = itemDone.Command;
+            if (command.CanExecute(null))
+                command.Execute(2);
+        }
+
+        private void menuItemChangeStatus(object sender, RoutedEventArgs e)
+        {
+            var command = ((MenuItem)sender).Command;
+            if (command.CanExecute(null))
+            {
+                command.Execute(true);
+                btnReloadReceiptUI.Command.Execute(null);
             }
         }
 
-        private void uscListOrder_Loaded(object sender, RoutedEventArgs e)
+        private void menuItemUpdate_Click(object sender, RoutedEventArgs e)
         {
-            viewControl.Name = "ListOrderUserControl";
-            viewControl.Content = new ListOrderUserControl();
-            btnDatHang.Opacity = 1;
-            btnNhapHang.Opacity = 0.8;
-            btnDSNhapHang.Opacity = 0.8;
-        }
-
-        private void btnNhapHang_Click(object sender, RoutedEventArgs e)
-        {
-            if (viewControl.Name != "ImportItemsUserControl")
+            var command = ((MenuItem)sender).Command;
+            if (command.CanExecute(null))
             {
-                viewControl.Content = new ImportItemsUserControl();
-                viewControl.Name = "ImportItemsUserControl";
+                int id;
+                if (Int32.TryParse(tbSelectedId.Text, out id))
+                {
+                    try
+                    {
+                        command.Execute(id);
+
+                        EditOrderWindow w = new EditOrderWindow();
+                        w.ShowDialog();
+                    }
+                    catch (Exception ex)
+                    {
+                        MessageBox.Show(ex.Message, "Không thể thực hiện thao tác này", MessageBoxButton.OK, MessageBoxImage.Error);
+                    }
+                }
             }
         }
 
-        private void btnDSNhapHang_Click(object sender, RoutedEventArgs e)
+        private void menuItemDelete_Click(object sender, RoutedEventArgs e)
         {
-            if (viewControl.Name != "ListImportedItemUserControl")
+            var result = MessageBox.Show("Xác nhận xóa đơn đặt hàng?", "Xác nhận", MessageBoxButton.OKCancel, MessageBoxImage.Question);
+            var command = ((MenuItem)sender).Command;
+
+            if (result == MessageBoxResult.OK && command.CanExecute(null))
             {
-                viewControl.Content = new ListImportedItemUserControl();
-                viewControl.Name = "ListImportedItemUserControl";
+                try
+                {
+                    command.Execute(true);
+
+                    MessageBox.Show("Xóa thành công!");
+                }
+                catch (Exception ex)
+                {
+                    MessageBox.Show(ex.Message, "Không thể thực hiện thao tác này", MessageBoxButton.OK, MessageBoxImage.Error);
+                }
             }
         }
+        private void ScrollViewer_PreviewMouseWheel(object sender, MouseWheelEventArgs e)
+        {
+            ScrollViewer scv = (ScrollViewer)sender;
+            scv.ScrollToVerticalOffset(scv.VerticalOffset - e.Delta);
+            e.Handled = true;
+        }
+
+        private void MenuItem_Click(object sender, RoutedEventArgs e)
+        {
+            EditOrderWindow w = new EditOrderWindow();
+            w.ShowDialog();
+        }
+
+        private void btnAddReceipt_Click(object sender, RoutedEventArgs e)
+        {
+            var result = MessageBox.Show("Xác nhận nhập hàng mới?", "Xác nhận", MessageBoxButton.OKCancel, MessageBoxImage.Question);
+            var command = ((Button)sender).Command;
+
+            if (result == MessageBoxResult.OK && command.CanExecute(null))
+            {
+                command.Execute(true);
+                MessageBox.Show("Nhập hàng thành công!");
+            }
+        }
+
     }
 }
